updated livestream ripper, remove last segment and merge output to mp4 and separate mp3

+++ livestream_rip.sh
@@ -20,14 +20,14 @@ targURL=$(curl -L $targ | tr "\'" "\n" |
 	ffmpeg -i $targURL -c copy -segment_time $timestamp -reset_timestamps 1 \
 	-f segment $outNAME%03d.mp4
 
-read -p "delete last segment? [y/n]: " opt
-if [[ $opt == "y" ]]; then
-    # clean up: since break is partial, remove last segment
-    rm $(ls $outNAME* | sort | tail -n 1);
+if (( $(ls act* | wc -l) > 1 )); then
+    read -p "delete last segment? [y/n]: " opt
+    if [[ $opt == "y" ]]; then
+        # clean up: since break is partial, remove last segment
+        rm $(ls $outNAME* | sort | tail -n 1);
+    fi
+    # merge segments into single file and create audio-only file for easy transcription
+    echo $(ls $outNAME*) | sed 's/ /\n/g' |  sed 's/^/file /g' > temp
+    ffmpeg -f concat -i temp -c copy "${outNAME}_all.mp4" && rm temp
+    ffmpeg -i "${outNAME}_all.mp4" -vn -ac 2 -b:a 192k "${outNAME}_all.mp3";
 fi
-
-# merge segments into single file and create audio-only file for easy transcription
-echo $(ls $outNAME*) | sed 's/ /\n/g' |  sed 's/^/file /g' > temp
-ffmpeg -f concat -i temp -c copy "${outNAME}_all.mp4" && rm temp
-ffmpeg -i "${outNAME}_all.mp4" -vn -ac 2 -b:a 192k "${outNAME}_all.mp3"
-
